+++ wolftpm/include.am
@@ -4,6 +4,8 @@
 
 nobase_include_HEADERS+= \
                          wolftpm/tpm2.h \
+                         wolftpm/tpm2_packet.h \
+                         wolftpm/tpm2_tis.h \
                          wolftpm/tpm2_wrap.h \
                          wolftpm/version.h \
                          wolftpm/visibility.h
